Fix compiler error: call of overloaded 'Pair(const char [9], time_t)' is ambiguous

+++ src/rpcmisc.cpp
@@ -1063,7 +1063,7 @@ UniValue getsnapshot(const UniValue& params, bool fHelp)
     }
     result = komodo_snapshot(top);
     if ( result.size() > 0 ) {
-        result.push_back(Pair("end_time", time(NULL)));
+        result.push_back(Pair("end_time", (int) time(NULL)));
     } else {
 	result.push_back(Pair("error", "no addressindex"));
     }
